Added multiple update options
Added the ability to update the admin status of a member, change the
price of an item and remove an item from the database.

While most of the code for these new options exist, I think there is a
problem with the change price and remove item, although this is probably
something minor I am over looking.